% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateFileDate}
\alias{calculateFileDate}
\title{Calulate File Date}
\usage{
calculateFileDate(date = Sys.Date(), dateFormat = "\%Y\%m\%d")
}
\arguments{
\item{date}{current date or date thta wants calculating for retrival}

\item{dateFormat}{format to reshape the date.}
}
\value{
Posix that has current date rearranged
}
\description{
Calulate File Date
}
