% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateURIforData}
\alias{generateURIforData}
\title{Generate the URI to request Data from}
\usage{

  generateURIforData(source = "http://ea-edubase-api-prod.azurewebsites.net/edubase",
  dataset = "edubasealldata", fileType = "csv",
  fileDate = calculateFileDate())
}
\arguments{
\item{source}{URI for Base Query}

\item{dataset}{Dataset that is requested}

\item{fileType}{the file type of the data}

\item{fileDate}{date of file to retrive.}
}
\value{
URI of file for retrival
}
\description{
Generate the URI to request Data from
}
