% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getSchoolInformation}
\alias{getSchoolInformation}
\title{Get SchoolInformation (Edubase)}
\usage{
getSchoolInformation(dataset = "All Data", refresh = FALSE, ...)
}
\arguments{
\item{dataset}{which Get Schools Information Data ("All Data", "All Data Links", "All Open State Schools")}

\item{refresh}{force a refresh of the data}

\item{...}{addtional parameters to pass into subsequent functions}
}
\value{
a tibble that has Edubase All Data in it
}
\description{
Access to the Edubase Dataset that describes all UK state schools and their Links/Relationships along with key statistics.
}
\examples{
myDataFrame <- getSchoolInformation()
}
